Move balance parsing logic to eth_blockchain
