+++ apps/eth_blockchain/test/support/dumb_adapter.ex
@@ -34,7 +34,7 @@ defmodule EthBlockchain.DumbAdapter do
   end
 
   def handle_call({:get_balances, _address, contract_addresses, _abi, _block}, _from, reg) do
-    balances = Map.new(contract_addresses, fn ca -> {ca, 123} end)
+    balances = Enum.into(contract_addresses, [], fn _ -> "0x7B" end)
     {:reply, {:ok, balances}, reg}
   end
